Revertimos el FDU673 por problemas que tuvo con otros modulos cuando pasamos el commercial_name a l10n_ec_niif_minimal (habia situaciones en las que no se ponia correctamente si no usabamos dicho modulo).
adicionalmente cambiamos un constructor de res_partner
