+++ ecua_partner/objects/res_company.py
@@ -33,9 +33,10 @@ class res_company(osv.Model):
         'avoid_duplicated_vat':fields.boolean('Avoid duplicated vat numbers', required=False,
                                               help='Evita la creación de contactos con el número de Cédula/RUC/Pasaporte duplicado',
                                               track_visibility='onchange'),
+        'comercial_name':fields.related('partner_id','comercial_name',type='char',relation='res.partner',string='Comercial Name',store=False, track_visibility='onchange'),
     }
     _defaults = {
-    'avoid_duplicated_vat': True, 
-}
+        'avoid_duplicated_vat': True,
+    }
 res_company()        
 # vim:expandtab:smartindent:tabstop=4:softtabstop=4:shiftwidth=4:
