+++ l10n_ec_niif_minimal/objects/__init__.py
@@ -22,4 +22,3 @@ import account_tax
 import account_invoice_tax
 import account_config
 import res_company
-import res_partner
