+++ l10n_ec_niif_minimal/objects/res_company.py
@@ -11,7 +11,6 @@ class res_company(osv.Model):
 
     _columns = {
         'special_tax_contributor_number':fields.char('Special Contributor Number', required=False, help='Special Contributors are designed by the Ecuadorian Tax Authority, the number is used in tax reporting', track_visibility='onchange'),
-        'comercial_name':fields.related('partner_id','comercial_name',type='char',relation='res.partner',string='Comercial Name',store=False, track_visibility='onchange'),
         'forced_accounting': fields.boolean('Forced to maintain accounting books', track_visibility='onchange')
     }
     _defaults = {
