+++ l10n_ec_niif_minimal/views/res_company_view.xml
@@ -11,9 +11,6 @@
 					<field name="special_tax_contributor_number"/>
 					<field name="forced_accounting"/>
 				</field>
-	            <field name="partner_id" position="after">
-	            	<field name="comercial_name"/>
-	            </field>
 	            <xpath expr="//form/sheet" position="after">
 		            <div class="oe_chatter">
                         <field name="message_follower_ids" widget="mail_followers"/>
