+++ l10n_ec_niif_minimal/views/res_partner_view.xml
@@ -1,17 +1,6 @@
 <?xml version="1.0" ?>
 <openerp>
     <data>
-		<record id="res_partner_inherit" model="ir.ui.view">
-			<field name="name">res.partner.inherit.comercial.name</field>
-			<field name="model">res.partner</field>
-			<field name="type">form</field>
-			<field name="inherit_id" ref="base_vat.view_partner_form"/>
-
-			<field name="arch" type="xml">
- 				<xpath expr="//field[@name='name']" position="after">
-					<field name="comercial_name" placeholder="Nombre Comercial..." attrs="{'invisible':[('is_company','=',False)]}"/>
-				</xpath>
-            </field>
-		</record>
+		<delete id="res_partner_inherit" model="ir.ui.view" />
    </data>
 </openerp>
